+++ querybuilder.go
@@ -1,3 +1,7 @@
+// Copyright 2020 The GoQueryBuilder Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package querybuilder
 
 import (
@@ -20,6 +24,12 @@ type Sqlbuilder struct {
 	limitStmt string
 	offsetStmt string
 	orderbyStmt string
+	// Insert table
+	insertStmt     string
+	// Update table
+	updateStmt     string
+	// Slice of columns, key value for updates & inserts.
+	columns        [][2]string
 	Dialect string //Can be postgres, mysql or (more to come)
 }
 
@@ -242,6 +252,16 @@ func (s *Sqlbuilder) Exists() string {
 
 func (s *Sqlbuilder) Build() string {
 
+	// insert
+	if s.insertStmt != `` {
+		return s.constructInsert()
+	}
+
+	// update
+	if s.updateStmt != `` {
+		return s.constructUpdate()
+	}
+
 	//build selects
 	if s.deletefromStmt == `` {
 		if s.selectStmt == `` {
@@ -291,7 +311,6 @@ func (s *Sqlbuilder) Build() string {
 	return returnString
 }
 
-
 func (s *Sqlbuilder) BuildInsert(table string, data interface{}, additionalQuery ...string) (string, error) {
 	dbCols, dbVals, err := mapStruct(data)
 	if err != nil {
@@ -307,7 +326,6 @@ func (s *Sqlbuilder) BuildInsert(table string, data interface{}, additionalQuery
 
 	return sql, nil
 }
-
 
 func (s *Sqlbuilder) BuildUpdate(table string, data interface{}, additionalQuery string) (string, error) {
 
@@ -337,148 +355,4 @@ func (s *Sqlbuilder) BuildUpdate(table string, data interface{}, additionalQuery
 	}
 
 	return sql, errors.New("sql build failed")
-}
-
-
-
-
-
-/**
-Helpers
-*/
-
-func (s *Sqlbuilder) formatSchema (schema string) string {
-	schemaParts := strings.Split(schema, ".")
-	finalSchemaStmt := ``
-
-	var dialectFormat string
-
-	switch strings.ToLower(s.Dialect) {
-	case "postgres":
-		dialectFormat = `"`
-		break
-	case "mysql":
-		dialectFormat = "`"
-		break
-	default:
-		dialectFormat = `"`
-	}
-
-
-	for _, v := range schemaParts {
-		if v == `*` {
-			finalSchemaStmt += `*`
-		}else{
-			part := strings.TrimSpace(v)
-			if string(part[0]) == dialectFormat && string(part[len(part) - 1]) == dialectFormat {
-				finalSchemaStmt += part + `.`
-			}else{
-				finalSchemaStmt += dialectFormat + part + dialectFormat + `.`
-			}
-
-		}
-	}
-
-	return strings.TrimSuffix(finalSchemaStmt, `.`)
-}
-
-func (s *Sqlbuilder) formatJoinOn (joinStmt string) string {
-	joinParts := strings.Split(joinStmt, "=")
-	finalJoinStmt := ``
-
-	for _, v := range joinParts {
-		finalJoinStmt += s.formatSchema(v) + ` = `
-	}
-
-	return strings.TrimSuffix(finalJoinStmt, ` = `)
-}
-
-var matchFirstCap = regexp.MustCompile("(.)([A-Z][a-z]+)")
-var matchAllCap   = regexp.MustCompile("([a-z0-9])([A-Z])")
-
-func toSnakeCase(str string) string {
-	snake := matchFirstCap.ReplaceAllString(str, "${1}_${2}")
-	snake  = matchAllCap.ReplaceAllString(snake, "${1}_${2}")
-	return strings.ToLower(snake)
-}
-
-func mapStruct (data interface{}) (dbCols []string, dbVals []string, error error){
-	fields := reflect.TypeOf(data)
-	values := reflect.ValueOf(data)
-
-	num := fields.NumField()
-
-	for i := 0; i < num; i++ {
-		field := fields.Field(i)
-		value := values.Field(i)
-
-		val, exists := field.Tag.Lookup("sqlb")
-
-		if exists {
-			dbCols = append(dbCols, "\"" + val + "\"")
-		}else{
-			dbCols = append(dbCols,  "\"" + toSnakeCase(field.Name)  + "\"")
-		}
-
-		var v string
-
-		v, ok := printInterface(value)
-		if !ok {
-			return dbCols, dbVals, errors.New("type: " + value.Kind().String() + " unsupported")
-		}
-		dbVals = append(dbVals, v)
-	}
-
-	return dbCols, dbVals, nil
-}
-
-func printInterface(value reflect.Value) (string, bool) {
-	var v string
-
-	switch value.Kind() {
-	case reflect.String:
-		v = "'" + sanitiseString(value.String()) + "'"
-	case reflect.Int:
-		v = strconv.FormatInt(value.Int(), 10)
-	case reflect.Int8:
-		v = strconv.FormatInt(value.Int(), 10)
-	case reflect.Int32:
-		v = strconv.FormatInt(value.Int(), 10)
-	case reflect.Int64:
-		v = strconv.FormatInt(value.Int(), 10)
-	case reflect.Float64:
-		v = fmt.Sprintf("%f", value.Float())
-	case reflect.Float32:
-		v = fmt.Sprintf("%f", value.Float())
-	case reflect.Bool:
-		if value.Bool(){
-			v = "TRUE"
-		}else{
-			v = "FALSE"
-		}
-	default:
-		return "", false
-	}
-
-	return v, true
-}
-
-func sanitiseString(str string) string {
-
-	if len(str) > 0 {
-		rebuildSingles := false
-
-		if string(str[0]) == "'" && string(str[len(str)-1]) == "'" {
-			rebuildSingles = true
-		}
-
-		str = strings.TrimSuffix(strings.TrimPrefix(str, "'"), "'")
-		str = strings.ReplaceAll(str, "'", "''")
-
-		if rebuildSingles {
-			str = "'" + str + "'"
-		}
-	}
-
-	return str
 }
